use class inheritance to add changes instead of making plot_2d even more messy
